TPs/TP2/code/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
certifi==2024.8.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via wandb
docker-pycreds==0.4.0
    # via wandb
einops==0.8.0
filelock==3.16.1
    # via
    #   torch
    #   triton
fsspec==2024.10.0
    # via torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
hydra-core==1.3.2
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.3
    # via torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
omegaconf==2.3.0
    # via hydra-core
packaging==24.2
    # via hydra-core
pillow==11.0.0
    # via torchvision
pip==24.2
platformdirs==4.3.6
    # via wandb
protobuf==5.29.0
    # via wandb
psutil==6.1.0
    # via wandb
pyyaml==6.0.2
    # via
    #   omegaconf
    #   wandb
requests==2.32.3
    # via wandb
sentry-sdk==2.19.0
    # via wandb
setproctitle==1.3.4
    # via wandb
setuptools==75.6.0
    # via
    #   torch
    #   wandb
six==1.16.0
    # via docker-pycreds
smmap==5.0.1
    # via gitdb
sympy==1.13.1
    # via torch
torch==2.5.1
    # via torchvision
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via torch
urllib3==2.2.3
    # via
    #   requests
    #   sentry-sdk
wandb==0.18.7
